// src/utils/types/types.ts
export enum ErrorMessages {
    emailFormat = "Incorrect email format",
    emailDomain = "Incorrect domain in email",
    emailDelimiter = "Missing @ symbol in email",
    hasSpaces = "Email should not contain spaces",
    nameEmpty = "Name cannot be empty",
    nameTooShort = "Name is too short",
    nameTooLong = "Name is too long",
    nameInvalidCharacters = "Name should only contain alphabetic characters, dashes, and spaces",
}